NutriLens-main/backend/requirements.txt
django
djangorestframework
django-cors-headers
openai
dotenv